Refactor for accel
